:recycle: Refactor seeding & add comments
